Django后端项目/renting/requirements.txt
asgiref==3.2.7
attr==0.3.1
attrs==19.3.0
autopep8==1.5.2
Django==3.0.6
django-cors-headers==3.3.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.6
prettytable==0.7.2
pycodestyle==2.6.0
pyecharts==1.8.1
pyflakes==2.2.0
PyJWT==1.7.1
pytz==2020.1
redis==3.5.3
simplejson==3.17.0
sqlparse==0.3.1
